bcmdhd: workaround broken AC video queue in IBSS mode
Re-apply this patch as it got lost with commit
6f16c5ddacd66f9122c6b74ce2979229ebb6be64 "tuna: Revert WiFi sleep power mode".
It still is necessary as can easily be tested with 'ping -Q 0xa IP'.

The BCM4330 has problems with the WME video (AC_VI) queue in IBSS mode. If we
put packets into this AC, the symptoms are excessive RTS/CTS, most packets get
lost and the device get's stuck and no more packets can be sent, even on other
queues.

Since the AC queue selection happens in the firmware based on the DSCP priority
and we don't know a way to configure this in the driver the workaround is to
simply downgrade all packets which would go into AC_VI (priority 4 and 5, TOS
0x80 and 0xa0) to AC_BE (priority 3 "Excellent Effort"). It's not pretty but
makes the device usable.

Change-Id: I833fb2f07dc5c46b823b70e93b9b05cc20a7b997
Signed-off-by: Bruno Randolf <br1@einfach.org>
Signed-off-by: Dan Trifan <jimsth@gmail.com>