omap2plus-cpufreq.c: only use maximum CPU overclock frequency if target frequency if above stock CPU maximum
If for example the highest CPU OPP (1.5 GHz) is enabled, all the below
frequencies are used as well, which increases the total number of
scaling steps and may hinder responsiveness. This patch only uses the
actual set maximum CPU frequency, if the requested frequency is above
the stock maximum frequency of 1.2 GHz.

Signed-off-by: Dan Trifan <jimsth@gmail.com>